requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
bottleneck==1.3.7
brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.112.2
fonttools==4.51.0
h11==0.14.0
idna==3.7
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.4
matplotlib==3.9.2
matplotlib-inline==0.1.7
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numexpr==2.8.7
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pymysql==1.0.2
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.9.0post0
pytz==2024.1
pyzmq==25.1.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.34
stack-data==0.6.2
starlette==0.38.2
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.1
unicodedata2
urllib3==2.2.3
uvicorn==0.20.0
wcwidth==0.2.13
wheel==0.44.0
zipp==3.20.2